backend/pyproject.toml
[tool.poetry]
name = "pour-priority-backend"
version = "0.1.0"
description = "FastAPI backend for Pour Priority project management app"
authors = ["Jordan <jordan@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlmodel = "^0.0.16"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.3.4"
python-multipart = "^0.0.9"
email-validator = "^2.1.1"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
httpx = "^0.27.0"
ruff = "^0.4.7"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
